Usando grid para que sea responsivo en celular. Usando un poco de foreach
