+++ src/index.html
@@ -6,6 +6,7 @@
     <link rel="stylesheet" href="style.css" />
   </head>
   <body>
+    <div id="bigDivs"></div>
     <div id="root">
       <p id="data1"></p>
       <p id="data2"></p>
